[TouchFW] Updated touch.c low power logic,L2x events no supported,eventlowpower cleanup - MH3-44884

+++ src/qtm_acq_samd21_0x0024_api.h
@@ -207,7 +207,7 @@ typedef struct
   qtm_acquisition_control_t* qtm_acq_control;
   uint16_t auto_scan_node_number;
   uint8_t auto_scan_node_threshold;
-  uint16_t auto_scan_trigger;  
+  uint8_t auto_scan_trigger;  
 }qtm_auto_scan_config_t;
